requirements.txt
fastapi
# fastapi-sessions
starlette
jinja2
webauthn
pyyaml
python-multipart
pyjwt
qrcode
Pillow
uvicorn[standard]
httpx